separated MRE home page from programQuestions component
